feat: Optionally collect contact information and hide seat numbers
These changes allow Paikkala to be used for e.g. Desucon's workshop programs, which require only a signup - in this case, it can be done by collecting contact information for the user and hiding the seat numbers in the ticket.

The only change (besides the obvious model changes and the corresponding migration) with an effect outside this use case: setting `Program.numbered_seats` automatically changes the behavior of `Program.reserve` as if it was called with `scatter=True`. The rationale for this is pretty obvious: for programs with unnumbered seats, seat numbering and thus scatter does not matter.